feat: relocate "game logic" in class "GameBoard"; add score counter
