+++ src/components/GameBoard/GameBoard.tsx
@@ -1,13 +1,24 @@
+/* eslint-disable class-methods-use-this */
+/* eslint-disable no-fallthrough */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable no-plusplus */
+/* eslint-disable no-alert */
+/* eslint-disable @typescript-eslint/no-unused-expressions */
+/* eslint-disable no-param-reassign */
+import { cloneDeep } from 'lodash';
 import React from 'react';
+import rotateMatrix from 'rotate-matrix';
 
-import { motion, randomCell } from '@/Logic/logicGame';
+import { Matrix, Props, State } from '@/components/GameBoard/GameBoard.model';
+import { Cell, ICell } from '@/components/GameBoard/GridCell.model';
 
 import NewGame from '../NewGame/NewGame';
 import Score from '../Score/Score';
 
-import { Props, State } from './GameBoard.model';
 import GridCell from './GridCell';
-import { Cell, ICell } from './GridCell.model';
 import styles from './style.scss';
 
 class GameBoard extends React.Component<Props, State> {
@@ -72,17 +83,168 @@ class GameBoard extends React.Component<Props, State> {
           },
         ],
       ],
+      score: 0,
     };
 
     this.newGame = this.newGame.bind(this);
+    this.randomRowIndex = this.randomRowIndex.bind(this);
+    this.randomColumnIndex = this.randomColumnIndex.bind(this);
+    this.randomCell = this.randomCell.bind(this);
+    this.calc = this.calc.bind(this);
+    this.removeEmptyCell = this.removeEmptyCell.bind(this);
+    this.motion = this.motion.bind(this);
+    this.scoreCounter = this.scoreCounter.bind(this);
   }
 
   componentDidMount() {
     document.addEventListener('keydown', (e:any) => {
       const { matrix } = this.state;
-      this.setState({ matrix: motion(e.key, matrix) });
+      this.setState({ matrix: this.motion(e.key, matrix) });
     });
     this.newGame();
+  }
+
+  randomRowIndex(motionType?: string): number {
+    let row = null;
+    switch (motionType) {
+      case 'ArrowDown': {
+        row = Math.floor(Math.random() * 3);
+        break;
+      }
+      case 'ArrowUp': {
+        row = Math.floor(Math.random() * 3) + 1;
+        break;
+      }
+      case 'ArrowLeft': {
+        row = Math.floor(Math.random() * 4);
+        break;
+      }
+      case 'ArrowRight': {
+        row = Math.floor(Math.random() * 4);
+        break;
+      }
+      default: row = Math.floor(Math.random() * 4);
+    }
+    return row;
+  }
+
+  randomColumnIndex(motionType?: string): number {
+    let column = null;
+    switch (motionType) {
+      case 'ArrowDown': {
+        column = Math.floor(Math.random() * 4);
+        break;
+      }
+      case 'ArrowUp': {
+        column = Math.floor(Math.random() * 4);
+        break;
+      }
+      case 'ArrowLeft': {
+        column = Math.floor(Math.random() * 3) + 1;
+        break;
+      }
+      case 'ArrowRight': {
+        column = Math.floor(Math.random() * 3);
+        break;
+      }
+      default: column = Math.floor(Math.random() * 4);
+    }
+    return column;
+  }
+
+  randomCell(matrix: Matrix, motionType?: string) {
+    if (matrix.flat().map(el => el?.value).includes(0)) {
+      const randomRow = this.randomRowIndex(motionType);
+      const randomColumn = this.randomColumnIndex(motionType);
+      matrix[randomRow][randomColumn]?.value === 0
+        ? matrix[randomRow][randomColumn]!.value = 2
+        : this.randomCell(matrix, motionType);
+    } else {
+      alert('YOU LOOOOOOOOOSE');
+    }
+  }
+
+  scoreCounter(increment: number) {
+    const { score } = this.state;
+    this.setState({ score: score + increment });
+  }
+
+  calc(arr: any) {
+    for (let i = arr.length - 1; i >= 0; i--) {
+      switch (true) {
+        case arr[i] === arr[i - 1]: {
+          arr[i] += arr[i - 1];
+          this.scoreCounter(arr[i]);
+          arr[i - 1] = 0;
+        }
+        case arr[i - 1] === 0: {
+          if (arr[i] === arr[i - 2]) {
+            arr[i] += arr[i - 2];
+            this.scoreCounter(arr[i]);
+            arr[i - 2] = 0;
+          }
+        }
+        case arr[i - 1] === 0 && arr[i - 2] === 0: {
+          if (arr[i] === arr[i - 3]) {
+            arr[i] += arr[i - 3];
+            this.scoreCounter(arr[i]);
+            arr[i - 3] = 0;
+          }
+        }
+        default: arr[i];
+      }
+    }
+    return arr;
+  }
+
+  removeEmptyCell(arr: Cell[]) {
+    for (let i = 0; i < arr.length - 1; i++) {
+      if (arr[i + 1]!.value === 0 && arr[i]!.value !== 0) {
+        const nextCell = cloneDeep(arr[i + 1]);
+        const currentCell = cloneDeep(arr[i]);
+        arr[i + 1] = currentCell;
+        arr[i] = nextCell;
+        i = 0;
+      }
+    }
+    return arr;
+  }
+
+  motion(motionType: string, matrix: Matrix) {
+    let rotateType: number;
+    switch (motionType) {
+      case 'ArrowDown': {
+        rotateType = -1;
+        break;
+      }
+      case 'ArrowUp': {
+        rotateType = 1;
+        break;
+      }
+      case 'ArrowLeft': {
+        rotateType = 2;
+        break;
+      }
+      case 'ArrowRight': {
+        rotateType = 4;
+        break;
+      }
+      default: rotateType = 4;
+    }
+    const rotationMatrix = rotateMatrix(matrix, rotateType);
+    const normalMatrix = rotateMatrix(rotationMatrix.map((cells: Cell[], rowIndex: number) => {
+      const amountCellValues = this.calc(cells.map((cell: Cell) => cell?.value));
+      return this.removeEmptyCell(
+        amountCellValues.map((cellValue: number, columnIndex: number) => ({
+          row: rowIndex,
+          column: columnIndex,
+          value: cellValue,
+        }))
+      );
+    }), -rotateType);
+
+    this.randomCell(normalMatrix, motionType);
+    return normalMatrix;
   }
 
   newGame() {
@@ -93,20 +255,20 @@ class GameBoard extends React.Component<Props, State> {
       id: el!.id,
       value: 0,
     })));
-    randomCell(newMatrix);
-    randomCell(newMatrix);
+    this.randomCell(newMatrix);
+    this.randomCell(newMatrix);
     this.setState({
       matrix: newMatrix,
     });
   }
 
   render() {
-    const { matrix } = this.state;
+    const { matrix, score } = this.state;
     return (
       <React.Fragment>
         <div className={styles['user-panel']}>
-          <NewGame newGame={this.newGame} />
-          <Score />
+          <NewGame newGame={() => this.newGame()} />
+          <Score score={score} />
         </div>
         <div className={styles['grid-container']}>
           {
