+++ src/components/Score/Score.tsx
@@ -1,26 +1,16 @@
 import React from 'react';
 
-import { State, Props } from './Score.model';
+import { Props } from './Score.model';
 import styles from './style.scss';
 
-class Score extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      score: 0,
-    };
-  }
-
-  render() {
-    const { score } = this.state;
-    return (
-      <React.Fragment>
-        <div>
-          Score: { score }
-        </div>
-      </React.Fragment>
-    );
-  }
-}
+const Score = ({ score }: Props): JSX.Element => (
+  <React.Fragment>
+    <div className={styles['score']}>
+      Score:
+      {' '}
+      { score }
+    </div>
+  </React.Fragment>
+);
 
 export default Score;
